\name{dtt-package}
\alias{dtt-package}
\docType{package}
\title{
Discrete Trigonometric Transforms
}
\description{
This package provides functions for 1D and 2D Discrete Cosine Transform (DCT), Discrete Sine Transform (DST) and Discrete Hartley Transform (DHT).
}
\details{
\tabular{ll}{
Package: \tab dtt\cr
Type: \tab Package\cr
Version: \tab 0.1-1\cr
Date: \tab 2007-02-25\cr
License: \tab GPL version 2 or newer.\cr
}
}
\author{
Maintainer: Lukasz Komsta <luke@novum.am.lublin.pl>
}
\keyword{ package }
